+++ imaudio/func/TriangleWave.cpp
@@ -1,7 +1,7 @@
 #include <math.h>
 #include "TriangleWave.h"
 
-long TriangleWave::sampleAtTime(double time) {
+double TriangleWave::sampleAtTime(double time) {
 	double slope = 2 * iWaveform::amplitude / 0.5;
 	time = time * iWaveform::frequency; // adjust time to frequency
 	time = time - (int)time; // remap to range [0, 1)
